==> src/73b_rum_mcmc_adl_4levels.R <==
library(rstan)
library(ggmcmc)

load("~/Data/JAHEAD/Process_Files/data_after_73a_74a.rda")
fit_model <- stan(file="src/73b_hierarchical_model_adl_4levels.stan", data=list_stan_adl, cores = 4, seed=1234,
                  iter = 100, warmup = 50, thin=1)
save(fit_model, file="~/Data/JAHEAD/Process_Files/stan_result_varying_intercept_adl_4levels.rda")
fit_model
# トレースプロット
stan_trace(fit_model, inc_warmup = TRUE)
stan_trace(fit_model)
# 事後分布
stan_dens(fit_model, separate_chains = TRUE)
# 自己相関
stan_ac(fit_model)
# R_hatの分布
stan_rhat(fit_model)
# 実効サンプルサイズ
stan_ess(fit_model)
